- Se valida si usuario está logueadio. - Linkeo modelos Post + User

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const path = require('path');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	const currentPage = req.query.page || 1;
@@ -54,19 +55,29 @@ exports.createPost = (req, res, next) => {
 	const imageUrl = req.file.path;
 	const title = req.body.title;
 	const content = req.body.content;
+	let creator;
 	const post = new Post({
 		title: title,
 		content: content,
 		imageUrl: imageUrl,
-		creator: { name: 'Victor Nilo' },
+		creator: req.userId,
 	});
 
 	post
 		.save()
 		.then(result => {
+			return User.findById(req.userId);
+		})
+		.then(user => {
+			creator = user;
+			user.posts.push(post);
+			return user.save();
+		})
+		.then(result => {
 			res.status(201).json( {
 				message: 'Post created successfully',
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name }
 			});
 		})
 		.catch( err => {
@@ -124,6 +135,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+			if (post.creator.toString() === req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -153,12 +169,22 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
-			// Check logged in user
+			if (post.creator.toString() === req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
 		.then( result => {
-			console.log(result);
+			return User.findById(req.user.userId);
+		})
+		.then(user => {
+			user.posts.pull(postId);
+			return user.save();
+		})
+		.then( result => {
 			res.status(200).json({message: 'Deleted post.'});
 		})
 		.catch( err => {
